remove inline styling wherever possible
