+++ src/components/Result/Table/TableView.css
@@ -1,0 +1,5 @@
+.result-table-container {
+    display: flex;
+    flex-direction: column;
+    overflow: auto;
+}
