+++ src/components/Result/Visualisation/Visualisation.css
@@ -1,0 +1,24 @@
+.visualisation {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    overflow-y: hidden;
+}
+
+.visualisation-box {
+    display: flex;
+    width: 50%;
+    border-right: 1px solid #dfdfdf;
+    border-top: 1px solid #dfdfdf;
+    border-bottom: 1px solid #dfdfdf;
+    overflow-y: hidden;
+}
+
+.visualisation-charts {
+    display: flex;
+    overflow-y: hidden;
+}
+
+.visualisation-dropdown {
+    display: flexs;
+}
